Add --maximum argument to run to limit number of processed entries
Add the argument `--maximum` to the `run` command which can be used to
specify how many new entries should be processed in each feed. This is
useful to prevent floods of emails in feeds that add many entries at
once, and also during development and debugging to be able to use real
feeds for testing without getting many emails every time.